zabezpieczenie przed usunieciem administratora

+++ Charity.Mvc/Controllers/AdminController.cs
@@ -79,12 +79,15 @@ namespace Charity.Mvc.Controllers
 
         [Authorize(Roles ="Admin")]
         public async Task<IActionResult> UserRemove(int id) 
-        {
+        {            
             var user = await _adminService.GetUserAsync(id);
             if (user == null) RedirectToAction("UserList", "Admin");
+            var adminRole = await RoleManager.FindByNameAsync("Admin");
+            var usersInAdminRole = _adminService.CountUsersInRoles(adminRole.Id);
+            if (usersInAdminRole <= 1 && await UserManager.IsInRoleAsync(user, "Admin") && await UserManager.GetUserAsync(User) == user)
+                    return RedirectToAction("UserList", "Admin");
             await _adminService.DeleteUserAsync(user);
-            return RedirectToAction("UserList", "Admin");
-            
+            return RedirectToAction("UserList", "Admin");            
         }
     }
 }
